feat(esxi): copy disk from image when creating vm
Before, we use delta disk based on vmdk in image_cache when creating vm.
In theory, creating a machine like this is faster and occupies less storage.
However, it may also affect performance and definitely cannot change the
size of the system disk.

Now, we copy disk from vmdk in image_cache to solve above questions.