fix(esxi): determine whether the disk is a root disk more reasonably
The order of the device is often sorted according to the size of the
key. The disk pointed to by the smallest key is not necessarily the
root disk or the first disk which often has the label of 'Hard disk 1'.
We try to find the label, if not, use the Key again to determine who
is the root disk.